+++ tests/clients/test_postgres.py
@@ -527,11 +527,11 @@ async def test_create_bulk_items_id_mismatch(
 
 async def test_db_setup_works_with_env_vars(api_client, database, monkeypatch):
     """Test that the application starts successfully if the POSTGRES_* environment variables are set"""
-    monkeypatch.setenv("POSTGRES_USER", database.user)
-    monkeypatch.setenv("POSTGRES_PASS", database.password)
-    monkeypatch.setenv("POSTGRES_HOST", database.host)
-    monkeypatch.setenv("POSTGRES_PORT", str(database.port))
-    monkeypatch.setenv("POSTGRES_DBNAME", database.dbname)
+    monkeypatch.setenv("PGUSER", database.user)
+    monkeypatch.setenv("PGPASSWORD", database.password)
+    monkeypatch.setenv("PGHOST", database.host)
+    monkeypatch.setenv("PGPORT", str(database.port))
+    monkeypatch.setenv("PGDATABASE", database.dbname)
 
     await connect_to_db(api_client.app)
     await close_db_connection(api_client.app)
@@ -564,11 +564,11 @@ class TestDbConnect:
         app fixture override to setup app with a customized db connection getter
         """
         postgres_settings = PostgresSettings(
-            postgres_user=database.user,
-            postgres_pass=database.password,
-            postgres_host=database.host,
-            postgres_port=database.port,
-            postgres_dbname=database.dbname,
+            pguser=database.user,
+            pgpassword=database.password,
+            pghost=database.host,
+            pgport=database.port,
+            pgdatabase=database.dbname,
         )
 
         logger.debug("Customizing app setup")