+++ tests/conftest.py
@@ -200,11 +200,11 @@ def api_client(request):
 @pytest.fixture(scope="function")
 async def app(api_client, database):
     postgres_settings = PostgresSettings(
-        postgres_user=database.user,
-        postgres_pass=database.password,
-        postgres_host=database.host,
-        postgres_port=database.port,
-        postgres_dbname=database.dbname,
+        pguser=database.user,
+        pgpassword=database.password,
+        pghost=database.host,
+        pgport=database.port,
+        pgdatabase=database.dbname,
     )
     logger.info("Creating app Fixture")
     time.time()
@@ -213,7 +213,6 @@ async def app(api_client, database):
         app,
         postgres_settings=postgres_settings,
         add_write_connection_pool=True,
-        write_postgres_settings=postgres_settings,
     )
 
     yield app
@@ -308,11 +307,11 @@ async def app_no_ext(database):
     )
 
     postgres_settings = PostgresSettings(
-        postgres_user=database.user,
-        postgres_pass=database.password,
-        postgres_host=database.host,
-        postgres_port=database.port,
-        postgres_dbname=database.dbname,
+        pguser=database.user,
+        pgpassword=database.password,
+        pghost=database.host,
+        pgport=database.port,
+        pgdatabase=database.dbname,
     )
     logger.info("Creating app Fixture")
     time.time()
@@ -320,7 +319,6 @@ async def app_no_ext(database):
         api_client_no_ext.app,
         postgres_settings=postgres_settings,
         add_write_connection_pool=True,
-        write_postgres_settings=postgres_settings,
     )
     yield api_client_no_ext.app
     await close_db_connection(api_client_no_ext.app)
@@ -349,11 +347,11 @@ async def app_no_transaction(database):
     )
 
     postgres_settings = PostgresSettings(
-        postgres_user=database.user,
-        postgres_pass=database.password,
-        postgres_host=database.host,
-        postgres_port=database.port,
-        postgres_dbname=database.dbname,
+        pguser=database.user,
+        pgpassword=database.password,
+        pghost=database.host,
+        pgport=database.port,
+        pgdatabase=database.dbname,
     )
     logger.info("Creating app Fixture")
     time.time()
@@ -380,11 +378,11 @@ async def app_client_no_transaction(app_no_transaction):
 @pytest.fixture(scope="function")
 async def default_app(database, monkeypatch):
     """Test default stac-fastapi-pgstac application."""
-    monkeypatch.setenv("POSTGRES_USER", database.user)
-    monkeypatch.setenv("POSTGRES_PASS", database.password)
-    monkeypatch.setenv("POSTGRES_HOST", database.host)
-    monkeypatch.setenv("POSTGRES_PORT", str(database.port))
-    monkeypatch.setenv("POSTGRES_DBNAME", database.dbname)
+    monkeypatch.setenv("PGUSER", database.user)
+    monkeypatch.setenv("PGPASSWORD", database.password)
+    monkeypatch.setenv("PGHOST", database.host)
+    monkeypatch.setenv("PGPORT", str(database.port))
+    monkeypatch.setenv("PGDATABASE", database.dbname)
     monkeypatch.delenv("ENABLED_EXTENSIONS", raising=False)
 
     monkeypatch.setenv("ENABLE_TRANSACTIONS_EXTENSIONS", "TRUE")